# Partial Loading PR3: Integrate 1) partial loading, 2) quantized models, 3) model patching